% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-info.R
\name{read_gpl_html}
\alias{read_gpl_html}
\title{read GPL information webpage}
\usage{
read_gpl_html(webpage)
}
\arguments{
\item{webpage}{passed on to \code{\link[xml2:read_xml]{xml2::read_html()}}.}
}
\value{
list
\enumerate{
\item info: tibble (\code{name}, \code{description})
\item sample: tibble, part of the platform table
}
}
\description{
read GPL information webpage
}
\examples{
read_gpl_html(system.file('extdata/GPL98.html', package = 'rGEO.data'))
read_gpl_html(system.file('extdata/GPL10400.html', package = 'rGEO.data'))

\dontrun{
    # feel free to run it, I don't do so to avoid accessing the Internet
    read_gpl_html('https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL16570')
}
}
