% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-info.R
\name{read_summary}
\alias{read_summary}
\title{read GEO datasets summary .txt file}
\usage{
read_summary(path)
}
\arguments{
\item{path}{string. path to the summary .txt file}
}
\value{
tibble
\enumerate{
\item accession
\item type
\item species
\item ftp
\item platform
\item data_type
\item title
\item description
}
}
\description{
refer to \href{https://github.com/dongzhuoer/thesis/r-lib/rGEO.data/blob/master/R-raw/data.Rmd#dataset}{here} for how to get the input file
}
\examples{
read_summary(system.file('extdata/gds_result-cut.txt', package = 'rGEO.data'))

}
